Makefile main.go sync.go
