+++ Makefile
@@ -83,6 +83,7 @@ fix: ${RTA}  # runs all linters and auto-fixes
 	go run tools/format_cmp_or/format_cmp_or.go
 	go run tools/format_unittests/format_unittests.go
 	go run tools/format_self/format_self.go
+	go run tools/sync-rta-version/main.go
 	make --no-print-directory keep-sorted
 	make --no-print-directory generate-json-schema
 	$(GOFUMPT) -l -w .
